Disabling trousers package for arc
Even though ARC gcc understands "-pie" option and attempts
to generate PIE binaries as of today PIE is not really supported
for user-space applications. Trousers configure scripts don't provide any
options which can disable pie usage. So there is no easy way to prevent
linking with -pie on ARC. The easiest solution is just to make
the trousers package not available on ARC.

Signed-off-by: Lada Trimasova <ltrimas@synopsys.com>
Cc: Alexey Brodkin <abrodkin@synopsys.com>
Cc: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Cc: Peter Korsgaard <peter@korsgaard.com>