qt5: enable JS and WebKit for ARC
This patch adds support of ARC architecture in Qt's JS core and Webkit.
Not only it allows to build everything that depends on
BR2_PACKAGE_QT5_JSCORE_AVAILABLE but built applications then really
work.

Unfortunately Qt 5.5 is obsolete now when 5.6 got released and
these patches won't be accepted in 5.5 branch I think.

Still with them here it will be possible to verify and fix all
issues that happen in run-time (if any) for 5.5 and once in
Buildroot Qt's version bumps to 5.6 I'll prepare new set of patches
and will upstream them.

Signed-off-by: Alexey Brodkin <abrodkin@synopsys.com>
Cc: Peter Seiderer <ps.report@gmx.net>
Cc: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Cc: Peter Korsgaard <peter@korsgaard.com>