ARC: Always compile for target with -mlong-calls
As of today GNU tools for ARC don't have linker relaxation implemented
and so we need to make sure full 32-bit offsets are used for calls.

Otherwise 25-bit (i.e. covering only =< 32Mb) offsets could be used
which will lead to link-time failures when linking huge apps or
libraries.

Very good example of this is Qt5WebKit. Even though it is now obsolete
when Qt5.5 gets built for ARC libQt5WebKit.so won't be linked
successfully.

Signed-off-by: Alexey Brodkin <abrodkin@synopsys.com>
Cc: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>