glibc: Install iconvconfig on target
In some cases we need to have gconv-modules.cache on target,
for example see [1] or glibc's test-suite (iconv/test-iconvconfig).

We do have iconvconfig installed in Buildroot's staging folder but
it gets never installed on targe because instead of calling glibc's
"install" target we simply copy a list of libs.

So then let's copy oven iconvconfig there as well in a similar manner.
Still we do it only if gconv libs are being copied otherwise it makes no
sense I guess.

[1] http://lists.busybox.net/pipermail/buildroot/2013-July/075097.html

Signed-off-by: Alexey Brodkin <abrodkin@synopsys.com>